requirements.txt
# Web框架和WebSocket支持
Flask==2.3.3
Flask-SocketIO==5.3.6

# AI API客户端
openai==1.3.5

# 环境变量管理
python-dotenv==1.0.0

# 异步支持
eventlet==0.33.3

# 生产环境服务器
gunicorn==21.2.0

# 数据库支持
sqlite3  # Python内置，无需安装

# 开发和测试工具
pytest==7.4.3
pytest-flask==1.3.0